+++ src/main/java/com/quasiris/qsf/json/EmptyNode.java
@@ -1,0 +1,4 @@
+package com.quasiris.qsf.json;
+
+public class EmptyNode {
+}