+++ src/main/java/com/quasiris/qsf/util/ElasticUtil.java
@@ -9,15 +9,26 @@ public class ElasticUtil {
     public static final String[] ELASTIC_RESERVED_CHARS = {"\\","+","-","=","&&","||",">","<","!","(",")","{","}","[","]","^","\"","~","*","?",":","/"," OR "," AND "};
     public static final String[] ELASTIC_ESCAPED_CHARS = {"\\\\","\\+","\\-","\\=","\\&&","\\||","\\>","\\<","\\!","\\(","\\)","\\{","\\}","\\[","\\]","\\^","\\\"","\\~","\\*","\\?","\\:","\\/"," \\OR "," \\AND "};
 
+    // https://www.elastic.co/guide/en/elasticsearch/reference/current/regexp-syntax.html
+    // . ? + * | { } [ ] ( ) " \
+    // # @ & < >  ~
+    public static final String[] ELASTIC_REGEX_RESERVED_CHARS = {"\\",".","?","+","*","|","{","}","[","]","(",")","\"","#","@","&","<",">","~"};
+    public static final String[] ELASTIC_REGEX_ESCAPED_CHARS = {"\\\\","\\.","\\?","\\+","\\*","\\|","\\{","\\}","\\[","\\]","\\(","\\)","\\\"","\\#","\\@","\\&","\\<","\\>","\\~"};
+
     public static String escape(String value) {
-        return escape(value, ELASTIC_RESERVED_CHARS);
+        return escape(value, ELASTIC_RESERVED_CHARS, ELASTIC_ESCAPED_CHARS);
     }
 
-    private static String escape(String value, String[] metaCharacters) {
+    public static String escapeRegex(String value) {
+        return escape(value, ELASTIC_REGEX_RESERVED_CHARS, ELASTIC_REGEX_ESCAPED_CHARS);
+    }
+
+
+    private static String escape(String value, String[] metaCharacters, String[] metaCharactersEscaped) {
         if(value != null) {
             for (int i = 0; i < metaCharacters.length; i++) {
                 if (value.contains(metaCharacters[i])) {
-                    value = value.replace(metaCharacters[i], ELASTIC_ESCAPED_CHARS[i]);
+                    value = value.replace(metaCharacters[i], metaCharactersEscaped[i]);
                 }
             }
         }