+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes qsf-integration project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- implement a logic to remove json nodes with the json substitutor
 
 ### Removed