+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes qsf-integration project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+
+### Removed
+
+### Changed
+
+### Deprecated
+
+### Fixed
+
+## [2.13]
+### Added
 - JsonBuilder: implement a logic to remove json nodes with the json substitutor
 - JsonBuilder: allow to replace multiple json keys
 - PipelineBuilder: allow to set an id for loop and condition filter for better debugging